requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.32.3
python-dotenv==1.0.1
